requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Caching==1.8.0
Flask-Login==0.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
